refactor: extend from the default theme instead of replacing it
Extend instead of duplicate theme css variables.
It seems more robust to not use the default theme but for development it's more clear to show what we are controlling ourselves vs relying on Tailwind defaults. Ensure visual regression testing is done.

Reasoning is that if Tailwind changes the defaults and the site is impacted we should probably check why the defaults have changed